feat: 홈 화면에 스캔 및 갤러리 진입 기능 추가 및 OCR 처리 연동 - VNDocumentCameraViewController를 활용한 문서 스캔 UI 연결 - Vision 프레임워크 기반 OCRService 구현 및 스캔/갤러리 이미지 OCR 처리 연동 - OCR 결과 콘솔 출력 처리 완료 - 레이아웃 및 버튼 크기 정렬 등 UI 세부 개선
